[GDPVal] Keep deliverable classification dependency-neutral
